fix: correct biome noise casting and monument generation
Correct integer casting in biomenoise from uint64 to int64
to properly handle negative parameter values and prevent overflow.

Update Ocean Monument generation logic to use getFeaturePos
for versions 1.18+ and getLargeStructurePos for earlier versions.

Add Stronghold structure type definition and validation.

修复生物群噪声类型转换及遗迹生成逻辑

修正生物群噪声中的整数转换，从 uint64 改为 int64，
以正确处理负参数值并防止溢出。

更新 Ocean Monument 生成逻辑，1.18+ 版本使用 getFeaturePos，
旧版本使用 getLargeStructurePos。

新增 Stronghold 结构类型定义及校验。

+++ biomenoise.go
@@ -147,16 +147,16 @@ func (bn *BiomeNoise) Sample(x, y, z int, flags uint32) int {
 	}
 
 	var np [6]uint64
-	np[NP_TEMPERATURE] = uint64(int32(bn.Climate[NP_TEMPERATURE].Sample(fx, 0, fz) * 10000.0))
-	np[NP_HUMIDITY] = uint64(int32(bn.Climate[NP_HUMIDITY].Sample(fx, 0, fz) * 10000.0))
-	np[NP_CONTINENTALNESS] = uint64(int32(bn.Climate[NP_CONTINENTALNESS].Sample(fx, 0, fz) * 10000.0))
-	np[NP_EROSION] = uint64(int32(bn.Climate[NP_EROSION].Sample(fx, 0, fz) * 10000.0))
-	np[NP_WEIRDNESS] = uint64(int32(bn.Climate[NP_WEIRDNESS].Sample(fx, 0, fz) * 10000.0))
+	np[NP_TEMPERATURE] = uint64(int64(int32(bn.Climate[NP_TEMPERATURE].Sample(fx, 0, fz) * 10000.0)))
+	np[NP_HUMIDITY] = uint64(int64(int32(bn.Climate[NP_HUMIDITY].Sample(fx, 0, fz) * 10000.0)))
+	np[NP_CONTINENTALNESS] = uint64(int64(int32(bn.Climate[NP_CONTINENTALNESS].Sample(fx, 0, fz) * 10000.0)))
+	np[NP_EROSION] = uint64(int64(int32(bn.Climate[NP_EROSION].Sample(fx, 0, fz) * 10000.0)))
+	np[NP_WEIRDNESS] = uint64(int64(int32(bn.Climate[NP_WEIRDNESS].Sample(fx, 0, fz) * 10000.0)))
 
 	// Depth calculation
 	if flags&SAMPLE_NO_DEPTH == 0 {
 		// TODO: Implement spline depth
-		np[NP_DEPTH] = uint64(int32((1.0 - float64(y*4)/128.0 - 83.0/160.0) * 10000.0))
+		np[NP_DEPTH] = uint64(int64(int32((1.0 - float64(y*4)/128.0 - 83.0/160.0) * 10000.0)))
 	} else {
 		np[NP_DEPTH] = 0
 	}
@@ -226,11 +226,11 @@ func get_np_dist(np []uint64, bt *BiomeTree, nodeIdx int) uint64 {
 
 	for i := 0; i < 6; i++ {
 		paramIdx := int((node >> (8 * i)) & 0xFF)
-		minVal := uint64(bt.Param[2*paramIdx+0])
-		maxVal := uint64(bt.Param[2*paramIdx+1])
+		minVal := int64(bt.Param[2*paramIdx+0])
+		maxVal := int64(bt.Param[2*paramIdx+1])
 
-		val := np[i]
-		var d uint64
+		val := int64(np[i])
+		var d int64
 		if val > maxVal {
 			d = val - maxVal
 		} else if val < minVal {
@@ -238,7 +238,7 @@ func get_np_dist(np []uint64, bt *BiomeTree, nodeIdx int) uint64 {
 		} else {
 			d = 0
 		}
-		ds += d * d
+		ds += uint64(d * d)
 	}
 	return ds
 }
